out/production/SSL-eKamba/requirement.txt
Automat	22.10.0	24.8.1
Bottleneck	1.3.7	1.4.0
CoLT5-attention	0.10.20	0.11.1
Jinja2	3.1.3	3.1.4
MarkupSafe	2.1.5	2.1.5
PyYAML	6.0.1	6.0.2
SQLAlchemy	2.0.29
Twisted	24.3.0	24.7.0
aiohttp	3.9.5	3.10.5
aiosignal	1.3.1	1.3.1
argparse	1.4.0	1.4.0
async-timeout	4.0.3	4.0.3
attrs	23.2.0	24.2.0
axial-positional-embedding	0.2.1	0.2.1
buildtools	1.0.6	1.0.6
causal-conv1d	1.2.0.post1
certifi	2022.12.7	2024.8.30
charset-normalizer	2.1.1	3.3.2
constantly	23.10.4	23.10.4
contourpy	1.2.0	1.3.0
cycler	0.11.0	0.12.1
docopt	0.6.2	0.6.2
einops	0.8.0	0.8.0
filelock	3.13.1	3.16.1
fonttools	4.51.0	4.53.1
frozenlist	1.4.1	1.4.1
fsspec	2024.2.0	2024.9.0
furl	2.1.3
greenlet	3.0.3	3.1.0
huggingface-hub	0.23.0	0.25.0
hyperlink	21.0.0	21.0.0
idna	3.4	3.10
incremental	22.10.0	24.7.2
joblib	1.4.2
kiwisolver	1.4.4	1.4.7
local-attention	1.9.1	1.9.15
mamba-ssm	1.2.1.post1	2.2.2
matplotlib	3.8.4	3.9.2
mkl-fft	1.3.8	1.3.8
mkl-random	1.2.4	1.2.4
mkl-service	2.4.0	2.4.1
mpld3	0.5.10	0.5.10
mpmath	1.3.0	1.3.0
multidict	6.0.5	6.1.0
networkx	3.2.1	3.3
ninja	1.11.1.1
numexpr	2.8.7	2.10.1
numpy	1.26.4	2.1.1
numpy	1.26.3	2.1.1
orderedmultidict	1.0.1	1.0.1
packaging	23.2	24.1
pandas	2.2.1
pillow	10.2.0	10.4.0
pillow	10.3.0	10.4.0
pip	23.3.1	24.2
ply	3.11	3.11
product-key-memory	0.2.10	0.2.11
psutil	5.9.8	6.0.0
pyparsing	3.0.9	3.1.4
pyparsing	3.1.2	3.1.4
python-dateutil	2.9.0.post0	2.9.0.post0
python-dateutil	2.8.2
pytz	2024.1	2024.2
redo	2.0.4	3.0.0
reformer-pytorch	1.4.4	1.4.4
regex	2024.4.28	2024.9.11
requests	2.28.1	2.32.3
safetensors	0.4.3	0.4.5
scikit-learn	1.4.2	1.5.2
scipy	1.13.0	1.14.1
setuptools	68.2.2	75.1.0
simplejson	3.19.2	3.19.3
sip	6.7.12	6.8.6
six	1.16.0	1.16.0
sympy	1.12	1.13.3
threadpoolctl	3.5.0	3.5.0
tokenizers	0.19.1	0.20.0
tomli	2.0.1	2.0.1
torch	2.1.1+cu118	2.4.1
torch_geometric	2.5.3	2.6.0
torchaudio	2.1.1+cu118	2.4.1
torchvision	0.16.1+cu118	0.19.1
tornado	6.3.3	6.4.1
tqdm	4.66.4	4.66.5
transformers	4.40.1	4.44.2
triton	2.1.0
typing_extensions	4.9.0	4.12.2
tzdata	2023.3	2024.1
unicodedata2	15.1.0	15.1.0
urllib3	1.26.13	2.2.3
wheel	0.43.0	0.44.0
yarl	1.9.4	1.11.1
zope.interface	6.3	7.0.3